+++ src/components/timeline/models.ts
@@ -1,4 +1,9 @@
-import {type ScheduledEvent} from "~/utils/types";
+export interface ScheduledEvent {
+    id: string,
+    name: string,
+    date: Date,
+    hours: number,
+}
 
 export interface DayViewProps {
     startDay: Date,
