--- src/utils/types.ts
@@ -1,7 +1,0 @@
-export interface ScheduledEvent {
-    id: string,
-    name: string, 
-    date: Date,
-    hours: number,
-}
-
